Wetzoek_GPUbackend/requirements.txt
acme==0.31.0
anyio==3.6.2
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certbot==0.27.0
certifi==2018.1.18
chardet==3.0.4
click==8.1.3
cloud-init==22.2
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.1.2
distlib==0.3.6
distro-info===0.18ubuntu0.18.04.1
dnspython==2.3.0
ec2-hibinit-agent==1.0.0
email-validator==1.3.1
fastapi==0.89.1
future==0.15.2
h11==0.14.0
hibagent==1.0.1
httpcore==0.16.3
httplib2==0.9.2
httptools==0.5.0
httpx==0.23.3
hyperlink==17.3.1
idna==3.4
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==2.1.2
matplotlib==2.1.1
mock==2.0.0
netifaces==0.10.4
nose==1.3.7
numpy==1.24.1
oauthlib==2.0.6
olefile==0.45.1
orjson==3.8.5
PAM==0.4.2
pandas==1.5.3
parsedatetime==2.4
pbr==3.1.1
pexpect==4.2.1
Pillow==5.1.0
pluggy==0.6.0
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pydantic==1.10.4
PyGObject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pyserial==3.4
pytest==3.3.2
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
pyxdg==0.25
PyYAML==6.0
requests==2.18.4
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rfc3986==1.5.0
scikit-learn==1.2.1
scipy==1.10.0
SecretStorage==2.3.1
service-identity==16.0.0
simplejson==3.13.2
six==1.11.0
sklearn==0.0.post2
sniffio==1.3.0
sos==4.3
ssh-import-id==5.7
starlette==0.22.0
systemd-python==234
threadpoolctl==3.1.0
Twisted==17.9.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.8
ufw==0.36
ujson==5.7.0
unattended-upgrades==0.1
urllib3==1.22
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
